+++ incubator/lnurl-java/lnurl-test/src/test/java/org/tbk/lnurl/test/LnurlAuthWalletUtilsTest.java
@@ -2,11 +2,9 @@ package org.tbk.lnurl.test;
 
 import fr.acinq.bitcoin.DeterministicWallet;
 import fr.acinq.bitcoin.DeterministicWallet.ExtendedPrivateKey;
-import fr.acinq.bitcoin.DeterministicWallet.KeyPath;
+import fr.acinq.bitcoin.KeyPath;
 import fr.acinq.secp256k1.Hex;
 import org.junit.jupiter.api.Test;
-import scala.collection.immutable.Seq;
-import scodec.bits.ByteVector;
 
 import java.net.URI;
 import java.security.SecureRandom;
@@ -21,7 +19,7 @@ class LnurlAuthWalletUtilsTest {
     @Test
     void itShouldCreateLinkingKeyForSiteWithRandomWallet() {
         var domainName = URI.create("https://example.com");
-        ExtendedPrivateKey masterPrivateKey = DeterministicWallet.generate(ByteVector.view(new SecureRandom().generateSeed(256)));
+        ExtendedPrivateKey masterPrivateKey = DeterministicWallet.generate(new SecureRandom().generateSeed(256));
 
         ExtendedPrivateKey linkingKey = LnurlAuthWalletUtils.deriveLinkingKey(masterPrivateKey, domainName);
         assertThat(linkingKey, is(notNullValue()));
@@ -42,7 +40,7 @@ class LnurlAuthWalletUtilsTest {
         var hashingKey = Hex.decode("7d417a6a5e9a6a4a879aeaba11a11838764c8fa2b959c242d43dea682b3e409b01");
 
         // expected path: m/138'/1603989739'/682320451'/2081389135'/1986563837'
-        KeyPath expectedKeyPath = new KeyPath(Seq.newBuilder().result())
+        KeyPath expectedKeyPath = new KeyPath("")
                 .derive(hardened(138L))
                 .derive(3751473387L)
                 .derive(2829804099L)
@@ -61,18 +59,18 @@ class LnurlAuthWalletUtilsTest {
         var walletSeed = Hex.decode("00".repeat(512));
 
         // expected path: m/138'/443735582/1653120675'/1066334360'/230068797'
-        KeyPath expectedKeyPath = new KeyPath(Seq.newBuilder().result())
+        KeyPath expectedKeyPath = new KeyPath("")
                 .derive(hardened(138L))
                 .derive(443735582L)
                 .derive(hardened(1653120675L))
                 .derive(hardened(1066334360L))
                 .derive(hardened(230068797L));
 
-        ExtendedPrivateKey masterPrivateKey = DeterministicWallet.generate(ByteVector.view(walletSeed));
+        ExtendedPrivateKey masterPrivateKey = DeterministicWallet.generate(walletSeed);
 
         ExtendedPrivateKey linkingKey = LnurlAuthWalletUtils.deriveLinkingKey(masterPrivateKey, domainName);
 
-        assertThat(linkingKey.path(), is(expectedKeyPath));
-        assertThat(linkingKey.path().toString(), is("m/138'/443735582/1653120675'/1066334360'/230068797'"));
+        assertThat(linkingKey.path, is(expectedKeyPath));
+        assertThat(linkingKey.path.toString(), is("m/138'/443735582/1653120675'/1066334360'/230068797'"));
     }
 }