+++ incubator/spring-lnurl/spring-lnurl-auth-example-application/src/integTest/java/org/tbk/lightning/lnurl/example/domain/WalletUserServiceImplIntegrationTest.java
@@ -1,5 +1,7 @@
 package org.tbk.lightning.lnurl.example.domain;
 
+import fr.acinq.bitcoin.PublicKey;
+import fr.acinq.secp256k1.Hex;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -24,9 +26,10 @@ class WalletUserServiceImplIntegrationTest {
     @Test
     @Transactional
     void itShouldCreateNewUserIfMissing() {
-        String validLinkingKeyHex = "0465d6177992064a24c24213230a0c3eeb5f2047c7286391c7ead608cda473f787af9afcae9af3a6a84f28a775ad257dbf6027448461455aaf482569237dda27bd";
+        String uncompressedValidLinkingKeyHex = "0465d6177992064a24c24213230a0c3eeb5f2047c7286391c7ead608cda473f787af9afcae9af3a6a84f28a775ad257dbf6027448461455aaf482569237dda27bd";
+        String compressedValidLinkingKeyHex = Hex.encode(PublicKey.compress(Hex.decode(uncompressedValidLinkingKeyHex)));
 
-        SimpleLinkingKey linkingKey = SimpleLinkingKey.fromHexStrict(validLinkingKeyHex);
+        SimpleLinkingKey linkingKey = SimpleLinkingKey.fromHex(compressedValidLinkingKeyHex);
 
         assertThat(walletUserService.findUser(linkingKey), is(Optional.empty()));
 
@@ -40,7 +43,7 @@ class WalletUserServiceImplIntegrationTest {
         assertThat(linkingKeys, hasSize(1));
 
         AuthLinkingKey authLinkingKey = linkingKeys.get(0);
-        assertThat(authLinkingKey.getLinkingKey(), is(validLinkingKeyHex));
+        assertThat(authLinkingKey.getLinkingKey(), is(compressedValidLinkingKeyHex));
         assertThat(authLinkingKey.toPublicKey().isValid(), is(true));
     }