+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Breaking
+- lnurl-auth: enforce compressed (33-byte) secp256k1 public key encoded as hex for `key` param
 - build: exclude protobuf files from jar file
 
 ### Changed
@@ -16,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - upgrade: update testcontainers from v1.15.2 to v1.16.2
 - upgrade: update grpc from v1.38.0 to v1.41.1
 - upgrade: update protobuf from v3.17.3 to v3.19.1
+- upgrade: switch from acinq/bitcoin-lib v0.19 to acinq/bitcoin-kmp v0.8.1
 
 ### Removed
 - module: sunset incubator module 'tbk-btcabuse-client'